+++ template/src/shared/contexts/UiComponentsWrapper.tsx
@@ -1,7 +1,8 @@
 import { ToastContainer } from '@/shared/contexts/toast';
 import React from 'react';
 import { DialogContainer } from './dialog';
-
+import { BottomSheetContainer } from './bottom-sheet';
+import { ContextMenuContainer } from './context-menu';
 const UiComponentsWrapper: React.FC<React.PropsWithChildren> = ({
   children,
 }) => {
@@ -9,6 +10,8 @@ const UiComponentsWrapper: React.FC<React.PropsWithChildren> = ({
     <>
       <DialogContainer />
       <ToastContainer />
+      <ContextMenuContainer />
+      <BottomSheetContainer />
       {children}
     </>
   );